Move SolutionButtonOption from Solution to Insight (#714)
