+++ src/Diagnostics.ModelsAndUtils/Models/ResponseExtensions/Solution.cs
@@ -81,11 +81,6 @@ namespace Diagnostics.ModelsAndUtils.Models.ResponseExtensions
         public SolutionTypeTag TypeTag { get; set; }
 
         /// <summary>
-        /// Options for the solution button UI look.
-        /// </summary>
-        public SolutionButtonOption ButtonOption { get; set; }
-
-        /// <summary>
         /// This is set automatically.
         /// Reads <see cref="OperationContext"/> to determine if an internal view should be rendered.
         /// </summary>
@@ -121,5 +116,35 @@ namespace Diagnostics.ModelsAndUtils.Models.ResponseExtensions
         {
             return Using(context.Resource.ResourceUri, actionOptions);
         }
+    }
+
+    public class SolutionButtonOption
+    {
+        public string Label { get; set; }
+
+        [JsonConverter(typeof(StringEnumConverter))]
+        public SolutionButtonType Type { get; set; } = SolutionButtonType.Button;
+
+        [JsonConverter(typeof(StringEnumConverter))]
+        public SolutionButtonPosition Position { get; set; } = SolutionButtonPosition.Bottom;
+
+        public SolutionButtonOption(string label, SolutionButtonType type = SolutionButtonType.Button, SolutionButtonPosition position = SolutionButtonPosition.Bottom)
+        {
+            this.Label = label;
+            this.Type = type;
+            this.Position = position;
+        }
+    }
+
+    public enum SolutionButtonType
+    {
+        Button,
+        Link,
+    }
+
+    public enum SolutionButtonPosition
+    {
+        Bottom = 0,
+        NextToTitle = 1,
     }
 }
